+++ tests/test_schema.py
@@ -1,5 +1,6 @@
 import pytest
-from .schemas import Schema_01
+from .schemas import (Schema_01, Schema_04)
+from confu.schema import (apply_default, apply_defaults)
 
 def test_schema_attributes():
     schema = Schema_01()
@@ -12,4 +13,35 @@ def test_schema_walk():
     def callback(attribute,path):
         paths.append(".".join(path))
     Schema_01().walk(callback)
-    assert sorted(paths) == sorted(['int_attr', 'list_attr', 'nested.int_attr', 'str_attr'])
+    assert sorted(paths) == sorted(['int_attr', 'list_attr', 'nested', 'nested.int_attr', 'str_attr'])
+
+
+@pytest.mark.parametrize("SchemaClass,config,expected", [
+    (Schema_04, {},
+        {"int_attr":123, "str_attr":"test", "list_attr":[],
+        "list_attr_w_default":[1,2,3], "nested":{"int_attr_choices":1}}),
+
+    (Schema_04, {"int_attr":999},
+        {"int_attr":999, "str_attr":"test", "list_attr":[],
+         "list_attr_w_default":[1,2,3], "nested":{"int_attr_choices":1}}),
+
+    (Schema_04, {"nested":{"int_attr":1}},
+        {"int_attr":123, "str_attr":"test", "list_attr":[],
+         "list_attr_w_default":[1,2,3],
+         "nested":{"int_attr_choices":1,"int_attr":1}}),
+
+    (Schema_04, {"nested":{"int_attr":1, "int_attr_choices":2}},
+        {"int_attr":123, "str_attr":"test", "list_attr":[],
+         "list_attr_w_default":[1,2,3],
+         "nested":{"int_attr_choices":2,"int_attr":1}}),
+
+    (Schema_04, {"list_attr_w_default":[4,5,6]},
+        {"int_attr":123, "str_attr":"test", "list_attr":[],
+         "list_attr_w_default":[4,5,6],
+         "nested":{"int_attr_choices":1}}),
+
+
+])
+def test_apply_defaults(SchemaClass, config, expected):
+    apply_defaults(Schema_04(), config)
+    assert expected == config